# Unmarshaller optimizations