@rjarry 2020-10-03
I think there may be an issue if the message header is not a multiple of 8 bytes long. There will be an additional read performed when reading the body since there is an alignment after reading the header. I forgot to take this into account. This will not break anything but this additional read could be avoided if the alignment bytes were prefetched from the start. I'll submit another PR with a fix.